HW10_Inference/PR1_Seldon/requirements.txt
# Seldon Core dependencies
seldon-core==1.17.0
# Вместо seldon-core-microservice используем установку через git
git+https://github.com/SeldonIO/seldon-core.git@release-1.17#egg=seldon-core-microservice&subdirectory=executor
torch==2.0.1
torchvision==0.15.2
numpy>=1.22.0
scikit-learn>=1.0.0
pillow>=9.0.0
seldon-core==1.17.0
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.6
requests>=2.28.0
# Дополнительные зависимости для тестирования и клиентского кода
numpy>=1.20.0
pandas>=1.3.0
torch>=1.9.0
torchvision>=0.10.0
Pillow>=8.0.0
requests>=2.25.0
pytest>=6.0.0
alibi==0.9.0
alibi-detect==0.10.5
torch>=1.13.0
torchvision>=0.14.0
numpy>=1.20.0
pillow>=9.0.0
scikit-learn>=1.0.0
seldon-core==1.17.0
# ML dependencies
torch>=2.2.0
torchvision>=0.17.0
pillow>=10.0.0
numpy>=1.26.0
pandas>=2.0.3
scikit-learn>=1.4.0
torch>=1.13.0
torchvision>=0.14.0
numpy>=1.20.0
pillow>=9.0.0
scikit-learn>=1.0.0
seldon-core==1.17.0
# API and deployment
fastapi>=0.95.0
uvicorn>=0.22.0
gunicorn>=21.2.0
pydantic>=2.0.0
torch>=1.13.0
torchvision>=0.14.0
numpy>=1.20.0
pillow>=9.0.0
scikit-learn>=1.0.0
seldon-core-microservice>=1.17.0
requests>=2.28.0
# Client dependencies
requests>=2.31.0
openapi-python-client>=0.14.0
torch==2.0.1
torchvision==0.15.2
numpy>=1.22.0
scikit-learn>=1.0.0
pillow>=9.0.0
seldon-core==1.17.0
fastapi>=0.95.0
uvicorn>=0.22.0
python-multipart>=0.0.6
requests>=2.28.0
# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
moto>=4.2.0
